Adding VS2017 support to samples

+++ Samples/ID@XboxSDK/Social/UWP/Cpp/Social.cpp
@@ -197,7 +197,7 @@ void Sample::RefreshUserList()
 
     for (auto x = 0; x < 15; x++)
     {
-        if (x < userList.size())
+        if (x < (int)userList.size())
         {
             users.push_back(std::make_shared<UserListItem>(userList.at(x)));
         }
